requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
greenlet==1.1.3.post0
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.4
pandas==1.5.1
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
django-environ